Move Api and Auth to lib
Now, it's not a microfronend
