+++ frontend/lib/auth/src/auth.js
