--- frontend/microfrontend/m_api/README.md
@@ -1,1 +1,0 @@
-Это билиотека для работы с Backend. По сути, реализация BFF.
