--- frontend/microfrontend/m_api/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "m_api",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "build": "rspack build --config rspack.config.js ",
-    "legacy:build": "webpack --config webpack.config.js --stats=minimal",
-    "serve": "serve dist -p 3010",
-    "start": "concurrently \"npm run build\" \"npm run serve\"",
-    "legacy:start": "concurrently \"npm run legacy:build\" \"npm run serve\""
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "devDependencies": {
-    "concurrently": "8.2.2",
-    "serve": "14.2.3",
-    "webpack": "5.92.1",
-    "webpack-cli": "5.1.4",
-    "@rspack/core": "0.7.5",
-    "@rspack/cli": "0.7.5",
-    "@rspack/dev-server": "0.7.5",
-    "@module-federation/enhanced": "0.2.5"
-  },
-  "dependencies": {
-  }
-}
