--- frontend/microfrontend/m_auth/.gitignore
@@ -1,2 +1,0 @@
-dist/*
-node_modules/*
