--- frontend/microfrontend/m_auth/README.md
@@ -1,1 +1,0 @@
-Это билиотека для работы с авотризации. По сути, реализация BFF.
