--- frontend/microfrontend/m_auth/rspack.config.js
@@ -1,61 +1,0 @@
-const {
-  container: { ModuleFederationPlugin },
-  HtmlRspackPlugin,
-} = require('@rspack/core');
-const path = require('path');
-module.exports = {
-  entry: './auth.js',
-  mode: 'development',
-  devtool: 'hidden-source-map',
-  output: {
-    publicPath: 'http://localhost:3012/',
-    clean: true,
-  },
-  experiments: {
-    css: true,
-  },
-  devServer: {
-    headers: {
-      "Access-Control-Allow-Origin": "*",
-      "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, PATCH, OPTIONS",
-      "Access-Control-Allow-Headers": "X-Requested-With, content-type, Authorization"
-    },
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(jpg|png|gif|jpeg)$/,
-        type: 'asset/resource',
-      },
-      {
-        test: /\.(js|jsx)$/,
-        use: {
-          loader: 'builtin:swc-loader',
-          options: {
-            jsc: {
-              parser: {
-                syntax: 'ecmascript',
-                jsx: true,
-              },
-              transform: {
-                react: {
-                  runtime: 'automatic',
-                },
-              },
-            },
-          },
-        },
-      },
-    ],
-  },
-
-  plugins: [
-    new ModuleFederationPlugin({
-      name: 'm_auth',
-      filename: 'remoteEntry.js',
-      exposes: {
-        "./auth" : "./auth.js"
-      },
-    }),
-  ],
-};
