+++ frontend/microfrontend/m_list/rspack.config.js
@@ -49,7 +49,6 @@ module.exports = {
       name: 'm_list',
       filename: 'remoteEntry.js',
       remotes: {
-        // 'm_lib': 'm_lib@http://localhost:3011/remoteEntry.js',
         'm_main': 'm_main@http://localhost:3000/remoteEntry.js',
       },
       exposes: {
