+++ frontend/microfrontend/m_register/package.json
@@ -20,17 +20,17 @@
   "devDependencies": {
     "@babel/core": "7.24.7",
     "@babel/preset-react": "7.24.7",
+    "@module-federation/enhanced": "0.2.5",
+    "@rspack/cli": "0.7.5",
+    "@rspack/core": "0.7.5",
+    "@rspack/dev-server": "0.7.5",
     "babel-loader": "9.1.3",
     "concurrently": "8.2.2",
     "css-loader": "7.1.2",
     "html-webpack-plugin": "5.6.0",
     "serve": "14.2.3",
     "webpack": "5.92.1",
-    "webpack-cli": "5.1.4",
-    "@rspack/core": "0.7.5",
-    "@rspack/cli": "0.7.5",
-    "@rspack/dev-server": "0.7.5",
-    "@module-federation/enhanced": "0.2.5"
+    "webpack-cli": "5.1.4"
   },
   "description": ""
 }
