+++ frontend/package.json
@@ -3,15 +3,17 @@
   "version": "1.0.0",
   "main": "webpack.config.js",
   "scripts": {
-    "build": "concurrently \"rspack build --config rspack.config.js\" \"cd microfrontend/m_api && rspack build --config rspack.config.js\" \"cd microfrontend/m_auth && rspack build --config rspack.config.js\" \"cd microfrontend/m_list && rspack build --config rspack.config.js\" \"cd microfrontend/m_popup && rspack build --config rspack.config.js\" \"cd microfrontend/m_register && rspack build --config rspack.config.js\"",
+    "build": "concurrently \"rspack build --config rspack.config.js\" \"cd lib/api && rspack build\" \"cd lib/auth && rspack build\" \"cd microfrontend/m_list && rspack build --config rspack.config.js\" \"cd microfrontend/m_popup && rspack build --config rspack.config.js\" \"cd microfrontend/m_register && rspack build\"",
     "legacy:build": "webpack --config webpack.config.js --stats=minimal",
     "serve": "serve dist -p 3000",
-    "start": "concurrently \"npm run serve\" \"cd microfrontend/m_api && npm run serve\" \"cd microfrontend/m_auth && npm run serve\" \"cd microfrontend/m_list && npm run serve\" \"cd microfrontend/m_popup && npm run serve\" \"cd microfrontend/m_register && npm run serve\"",
+    "start": "concurrently \"npm run serve\" \"cd microfrontend/m_list && npm run serve\" \"cd microfrontend/m_popup && npm run serve\" \"cd microfrontend/m_register && npm run serve\"",
     "legacy:start": "concurrently \"npm run legacy:build\" \"npm run serve\"",
     "test:microfrontend": "npx jest ./index.spec.js",
     "test": "npm test"
   },
   "dependencies": {
+    "api": "link:lib/api",
+    "auth": "link:lib/auth",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-router-dom": "^5.2.0"
@@ -23,21 +25,21 @@
     "@babel/core": "7.24.7",
     "@babel/preset-react": "7.24.7",
     "@module-federation/enhanced": "0.2.5",
+    "@playwright/test": "^1.45.0",
     "@rspack/cli": "0.7.5",
     "@rspack/core": "0.7.5",
     "@rspack/dev-server": "0.7.5",
+    "@types/node": "^20.14.9",
     "babel-loader": "9.1.3",
     "concurrently": "^8.2.2",
     "css-loader": "7.1.2",
+    "dotenv": "^16.0.2",
     "html-webpack-plugin": "5.6.0",
     "react-singleton-context": "^1.0.5",
     "serve": "14.2.3",
+    "wait-port": "^1.1.0",
     "webpack": "5.92.1",
-    "webpack-cli": "5.1.4",
-    "@playwright/test": "^1.45.0",
-    "@types/node": "^20.14.9",
-    "dotenv": "^16.0.2",
-    "wait-port": "^1.1.0"
+    "webpack-cli": "5.1.4"
   },
   "eslintConfig": {
     "extends": [
