+++ frontend/rspack.config.js
@@ -50,9 +50,6 @@ module.exports = {
       name: 'm_main',
       filename: 'remoteEntry.js',
       remotes: {
-        'm_api': 'm_api@http://localhost:3010/remoteEntry.js',
-        'm_lib': 'm_lib@http://localhost:3011/remoteEntry.js',
-        'm_auth': 'm_auth@http://localhost:3012/remoteEntry.js',
         'm_list': 'm_list@http://localhost:3013/remoteEntry.js',
         'm_register': 'm_register@http://localhost:3014/remoteEntry.js',
         'm_popup': 'm_popup@http://localhost:3015/remoteEntry.js',
