+++ frontend/src/components/App.js
@@ -13,8 +13,8 @@ import Register from "m_register/Register";
 import Login from "m_register/Login";
 import InfoTooltip from "m_popup/InfoTooltip";
 import ProtectedRoute from "./ProtectedRoute";
-import api from "m_api/api";
-import * as auth from "m_auth/auth";
+import api from "api";
+import * as auth from "auth";
 import { AddPlacePopupContext } from "m_popup/AddPlacePopupContext";
 import { EditAvatarPopupContext } from "m_popup/EditAvatarPopupContext";
 import { EditProfilePopupContext } from "m_popup/EditProfilePopupContext";
